move: move shared or global config files to view/common
view 설정 파일 같아서 view 폴더 밑에 둠.
