+++ app/src/main/java/com/nicer/attiary/view/common/GlobalApplication.kt
@@ -1,9 +1,10 @@
-package com.nicer.attiary
+package com.nicer.attiary.view.common
 
 import android.app.Application
 import android.content.Context
 import androidx.lifecycle.ProcessLifecycleOwner
 import com.kakao.sdk.common.KakaoSdk
+import com.nicer.attiary.R
 
 class GlobalApplication : Application() {
 	companion object {
